Extracted Bolgar score components into seperate components
